Use GitHub Actions CI for tests
This commit also temporarily comments out flake8 in pre-install
as there are still legacy linting issues to deal with in a general
refactor.

Siegfried installation was failing due to the sunsetting of
bintray, which was hosting the deb package. See:
https://github.com/richardlehane/siegfried/issues/163

For now the CI config has been modified to install siegfried via
homebrew instead, which is newly supported on Linux.